backend/app/auth.py
from datetime import datetime, timedelta, timezone

from fastapi import Depends, HTTPException
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError, jwt
from passlib.context import CryptContext
from sqlalchemy import select

from app.config import ACCESS_TOKEN_EXPIRE_MINUTES, JWT_ALGORITHM, JWT_SECRET
from app.database import get_db
from app.models import User

pwd_context = CryptContext(schemes=["bcrypt"])

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="login")


def hash_password(password: str) -> str:
    return pwd_context.hash(password)


def verify_password(password: str, hash_password):
    return pwd_context.verify(password, hash_password)


def create_token(user_id):
    data = {
        "sub": str(user_id),
        "exp": datetime.now(timezone.utc) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES),
    }
    token = jwt.encode(data, JWT_SECRET, algorithm=JWT_ALGORITHM)
    return token


async def get_current_user(token: str = Depends(oauth2_scheme), db=Depends(get_db)):
    try:
        data = jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALGORITHM])
        user_id = int(data["sub"])
    except JWTError:
        raise HTTPException(status_code=401, detail="Invalid or expired token")

    sql_result = await db.execute(select(User).where(User.id == user_id))

    user = sql_result.scalar_one_or_none()
    if user is None:
        raise HTTPException(status_code=401, detail="User not found")
    return user
