backend/pyproject.toml
[project]
name = "HiveNet"
version = "0.1.0"
description = "Social media backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "sqlalchemy",
    "asyncpg",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "uvicorn",
    "pydantic[email]",
    "bcrypt==4.0.1",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
